pyproject.toml
[project]
name = "archivebox"
version = "0.8.5rc44"
requires-python = ">=3.10"
description = "Self-hosted internet archiving solution."
authors = [{name = "Nick Sweeting", email = "pyproject.toml@archivebox.io"}]
license = {text = "MIT"}
readme = "README.md"
keywords = ["internet archiving", "web archiving", "digipres", "warc", "preservation", "backups", "archiving", "web", "bookmarks", "puppeteer", "browser", "download"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: Sociology :: History",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Archiving",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: System :: Recovery Tools",
    "Topic :: Utilities",
    "Typing :: Typed",
]


dependencies = [
    # ... archivebox/vendor/*             # see vendored libs here
    ############# Django / Core Libraries #############
    "setuptools>=74.1.0",
    "django>=5.1.1,<6.0",
    "django-ninja>=1.3.0",
    "django-extensions>=3.2.3",
    "mypy-extensions>=1.0.0",
    "channels[daphne]>=4.1.0",
    "django-signal-webhooks>=0.3.0",
    "django-admin-data-views>=0.4.1",
    "django-object-actions>=4.3.0",
    "django-charid-field>=0.4",
    "django-pydantic-field>=0.3.10",
    "django-jsonform>=2.22.0",
    "django-stubs>=5.0.4",
    "django-huey>=1.2.1",
    "django-huey-monitor>=0.9.0",
    ############# Python Helper Libraries ############
    "pluggy>=1.5.0",
    "requests>=2.32.3",
    "dateparser>=1.2.0",
    "feedparser>=6.0.11",
    "w3lib>=2.2.1",
    "rich>=13.8.0",
    "rich-argparse>=1.5.2",
    "ulid-py>=1.1.0",
    "typeid-python>=0.3.1",
    "psutil>=6.0.0",
    "supervisor>=4.2.5",
    "python-crontab>=3.2.0",          # for: archivebox schedule
    "croniter>=3.0.3",                # for: archivebox schedule
    "ipython>=8.27.0",                # for: archivebox shell
    "py-machineid>=0.6.0",            # for: machine/detect.py calculating machine guid
    "python-benedict[io,parse]>=0.33.2",
    "pydantic-settings>=2.5.2",
    "atomicwrites==1.4.1",
    "django-taggit==1.3.0",
    "base32-crockford==0.3.0",
    # "pocket@git+https://github.com/tapanpandita/pocket.git@v0.3.7",
    "pydantic-pkgr>=0.5.3",
    ############# Plugin Dependencies ################
    "sonic-client>=1.0.0",
    "yt-dlp>=2024.8.6",               # for: media"
]

[project.optional-dependencies]
sonic = [
    # echo "deb [signed-by=/usr/share/keyrings/valeriansaliou_sonic.gpg] https://packagecloud.io/valeriansaliou/sonic/debian/ bookworm main" > /etc/apt/sources.list.d/valeriansaliou_sonic.list
    # curl -fsSL https://packagecloud.io/valeriansaliou/sonic/gpgkey | gpg --dearmor -o /usr/share/keyrings/valeriansaliou_sonic.gpg
    # apt install sonic
    # no-op, now included by default:
    # "sonic-client>=1.0.0",
]
ldap = [
    # apt install libsasl2-dev python-dev-is-python3 libldap2-dev libssl-dev
    "python-ldap>=3.4.3",
    "django-auth-ldap>=4.1.0",
]
all = [
    "archivebox[sonic,ldap]"
]

[tool.uv]
dev-dependencies = [
    ### BUILD
    "uv",
    "pip>=24.2",
    "setuptools>=75.1.0",
    "wheel>=0.44.0",
    "homebrew-pypi-poet>=0.10.0",      # for: generating archivebox.rb brewfile list of python packages
    ### DOCS
    "recommonmark>=0.7.1",
    "sphinx",
    "sphinx-rtd-theme>=2.0.0",
    ### DEBUGGING
    "django-debug-toolbar>=4.4.6",
    "djdt_flamegraph>=0.2.13",
    "ipdb>=0.13.13",
    "requests-tracker>=0.3.3",
    "logfire[django]>=0.51.0",
    "opentelemetry-instrumentation-django>=0.47b0",
    "opentelemetry-instrumentation-sqlite3>=0.47b0",
    "viztracer",                                     # usage: viztracer ../.venv/bin/archivebox manage check
    # "snakeviz",                                      # usage: python -m cProfile -o flamegraph.prof ../.venv/bin/archivebox manage check
    ### TESTING
    "pytest>=8.3.3",
    "bottle>=0.13.1",
    ### LINTING
    "ruff>=0.6.6",
    "flake8>=7.1.1",
    "mypy>=1.11.2",
    "django-autotyping>=0.5.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
# https://github.com/astral-sh/uv/issues/3957

[tool.setuptools]
packages = ["archivebox"]
package-dir = {"archivebox" = "archivebox"}

[tool.ruff]
line-length = 140
target-version = "py310"
src = ["archivebox"]
exclude = ["*.pyi", "typings/", "migrations/", "vendor/"]

# https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
ignore = ["E731", "E303", "E266", "E241", "E222"]

[tool.pytest.ini_options]
testpaths = [ "tests" ]

[tool.mypy]
mypy_path = "archivebox,archivebox/typings"
namespace_packages = true
explicit_package_bases = true
# follow_imports = "silent"
# ignore_missing_imports = true
# disallow_incomplete_defs = true
# disallow_untyped_defs = true
# disallow_untyped_decorators = true
# exclude = "tests/.*"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[tool.pyright]
include = [
    "archivebox",
]
exclude = [
    ".venv",
    "**/*.pyi",
    "**/__init__.pyi",
    "**/node_modules",
    "**/__pycache__",
    "**/migrations",
    "archivebox/vendor",
]
stubPath = "./archivebox/typings"
venvPath = "."
venv = ".venv"
# ignore = ["src/oldstuff"]
# defineConstant = { DEBUG = true }
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.10"
pythonPlatform = "Linux"


[project.scripts]
archivebox = "archivebox.cli:main"


[project.urls]
Homepage = "https://github.com/ArchiveBox/ArchiveBox"
Source = "https://github.com/ArchiveBox/ArchiveBox"
Documentation = "https://github.com/ArchiveBox/ArchiveBox/wiki"
"Bug Tracker" = "https://github.com/ArchiveBox/ArchiveBox/issues"
Changelog = "https://github.com/ArchiveBox/ArchiveBox/releases"
Roadmap = "https://github.com/ArchiveBox/ArchiveBox/wiki/Roadmap"
Community = "https://github.com/ArchiveBox/ArchiveBox/wiki/Web-Archiving-Community"
Demo = "https://demo.archivebox.io"
Donate = "https://github.com/ArchiveBox/ArchiveBox/wiki/Donations"


